Make variable grammar dependant
This makes variables scoped by grammar. Previously, you could set a
variable in one grammar, and access it from any grammar, as long as the
rule case that defined it was executed before hand. Now, only cases in
that grammar can use it.

There are still issues to do with the fact that variables can be used at
any time after they've been defined, even after their introductory rule
returned. I'm not certain that this is enough of an issue to bother
trying to fix it though.

The next step is an autovivify pragma for rules to solve several corner
cases, with the main one being wanting to use exhaustible variables from
prefix/suffix pragmas.