Refactor front-end error-handling
This refactors the front-end to use a tree for capturing errors, instead
of throwing exceptions.

This has the benefit that you will receive notifications about all of
the error messages you have, instead of only the first.

I'm a bit fuzzy on the details, since it's been a while since I wrote
these changes.